+++ libft/ft_putnbr_fd.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_putnbr_fd(int n, int fd)
+int	ft_putnbr_fd(long long n, int fd)
 {
 	int	i;
